+++ crates/sui-bridge-indexer/Cargo.toml
@@ -43,10 +43,9 @@ sui-test-transaction-builder.workspace = true
 test-cluster.workspace = true
 hex-literal = "0.3.4"
 sui-indexer.workspace = true
-sui-bridge-test-utils.workspace = true
 diesel_migrations = "2.2.0"
 sui-indexer-builder = { workspace = true, features = ["test-utils"] }
-move-core-types.workspace = true
+sui-bridge = { workspace = true, features = ["test-utils"] }
 
 [[bin]]
 name = "bridge-indexer"
