+++ crates/sui-bridge/Cargo.toml
@@ -48,12 +48,15 @@ mysten-common.workspace = true
 enum_dispatch.workspace = true
 sui-json-rpc-api.workspace = true
 sui-test-transaction-builder.workspace = true
+hex-literal = { version = "0.3.4", optional = true }
+test-cluster = { workspace = true, optional = true }
 
 [dev-dependencies]
 sui-types = { workspace = true, features = ["test-utils"] }
 sui-json-rpc-types = { workspace = true, features = ["test-utils"] }
 sui-config.workspace = true
 sui-test-transaction-builder.workspace = true
-test-cluster.workspace = true
-hex-literal = "0.3.4"
 maplit = "1.0.2"
+
+[features]
+test-utils = ["hex-literal", "test-cluster"]
