+++ crates/sui-bridge/src/e2e_tests/complex.rs
@@ -2,9 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::client::bridge_authority_aggregator::BridgeAuthorityAggregator;
-use crate::e2e_tests::basic::initiate_bridge_eth_to_sui;
-use crate::e2e_tests::basic::initiate_bridge_sui_to_eth;
-use crate::e2e_tests::test_utils::BridgeTestClusterBuilder;
+
+use crate::e2e_tests::test_utils::{
+    initiate_bridge_eth_to_sui, initiate_bridge_sui_to_eth, BridgeTestClusterBuilder,
+};
 use crate::sui_transaction_builder::build_sui_transaction;
 use crate::types::{BridgeAction, EmergencyAction};
 use crate::types::{BridgeActionStatus, EmergencyActionType};
