+++ crates/sui-bridge/src/e2e_tests/mod.rs
@@ -1,6 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+#[cfg(test)]
 mod basic;
+#[cfg(test)]
 mod complex;
 pub mod test_utils;
